+++ evals/benchmarks/mcpmark/__init__.py
@@ -1,0 +1,1 @@
+"""MCPMark adapter and paired-run entrypoints."""